improve(serverless): refactor scripts into their own package & run in CI (#2186)
