+++ packages/core/package.json
@@ -32,7 +32,6 @@
     "moment": "^2.24.0",
     "node-fetch": "^2.3.0",
     "solidity-coverage": "^0.7.2",
-    "supertest": "^4.0.2",
     "truffle": "^5.1.16",
     "truffle-assertions": "^0.9.2",
     "truffle-deploy-registry": "^0.5.1",
