+++ packages/serverless-orchestration/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require("./src/ServerlessHub"),
+  ...require("./src/ServerlessSpoke")
+};
