+++ packages/serverless-orchestration/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@uma/serverless-orchestration",
+  "version": "1.1.0",
+  "description": "Serverless orchestration scripts to run a parallel instances of scripts or bots",
+  "dependencies": {
+    "@truffle/contract": "^4.2.20",
+    "@uma/common": "^1.0.1",
+    "@uma/financial-templates-lib": "^1.2.0"
+  },
+  "devDependencies": {
+    "@awaitjs/express": "^0.3.0",
+    "@openzeppelin/contracts": "3.0.0",
+    "chai": "^4.2.0",
+    "coveralls": "^3.1.0",
+    "ethereumjs-wallet": "^1.0.0",
+    "express": "^4.17.1",
+    "minimist": "^1.2.0",
+    "mocha": "^8.1.2",
+    "moment": "^2.24.0",
+    "node-fetch": "^2.3.0",
+    "supertest": "^4.0.2",
+    "truffle": "^5.1.16",
+    "truffle-assertions": "^0.9.2",
+    "web3": "1.2.7",
+    "@uma/core": "npm:@uma/core@1.2.0"
+  },
+  "homepage": "https://umaproject.org",
+  "license": "AGPL-3.0-or-later",
+  "publishConfig": {
+    "registry": "https://registry.npmjs.com/",
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/UMAprotocol/protocol.git"
+  },
+  "main": "index.js",
+  "files": [
+    "/src/**/*.js"
+  ],
+  "scripts": {
+    "test": "yarn truffle test test test/*.js"
+  },
+  "bugs": {
+    "url": "https://github.com/UMAprotocol/protocol/issues"
+  }
+}
